+++ antareslauncher/use_cases/retrieve/study_retriever.py
@@ -1,14 +1,8 @@
 from antareslauncher.data_repo.data_reporter import DataReporter
 from antareslauncher.study_dto import StudyDTO
-from antareslauncher.use_cases.retrieve.clean_remote_server import (
-    RemoteServerCleaner,
-)
-from antareslauncher.use_cases.retrieve.download_final_zip import (
-    FinalZipDownloader,
-)
-from antareslauncher.use_cases.retrieve.final_zip_extractor import (
-    FinalZipExtractor,
-)
+from antareslauncher.use_cases.retrieve.clean_remote_server import RemoteServerCleaner
+from antareslauncher.use_cases.retrieve.download_final_zip import FinalZipDownloader
+from antareslauncher.use_cases.retrieve.final_zip_extractor import FinalZipExtractor
 from antareslauncher.use_cases.retrieve.log_downloader import LogDownloader
 from antareslauncher.use_cases.retrieve.state_updater import StateUpdater